complete double-tetris game on one pc

+++ double-tetris/main.js
@@ -2,13 +2,11 @@ const playerElements = document.querySelectorAll('.player');
 const tetrises = [];
 
 [...playerElements].forEach(element => {
-  const canvas = element.querySelector('.tetris');
-  const tetris = new Tetris(canvas);
+  const tetris = new Tetris(element);
   tetrises.push(tetris);
 });
 
-
-document.addEventListener('keydown', e => {
+const keyListeners = e => {
 
   [
     [65, 68, 83, 81, 69],
@@ -16,30 +14,39 @@ document.addEventListener('keydown', e => {
   ].forEach((key, index) => {
 
     const player = tetrises[index].player;
+    if (e.type == 'keydown') {
+      // left
+      if (e.keyCode == key[0]) {
+        player.move(-1);
+      }
 
-    // left
-    if (e.keyCode == key[0]) {
-      player.move(-1);
-    }
+      // right
+      if (e.keyCode == key[1]) {
+        player.move(1);
+      }
 
-    // right
-    if (e.keyCode == key[1]) {
-      player.move(1);
+      // rotate 1
+      if (e.keyCode == key[3]) {
+        player.rotate(-1);
+      }
+
+      // rotate 2
+      if (e.keyCode == key[4]) {
+        player.rotate(1);
+      }
     }
 
     // down
     if (e.keyCode == key[2]) {
-      player.drop();
-    }
-
-    // rotate 1
-    if (e.keyCode == key[3]) {
-      player.rotate(-1);
-    }
-
-    // rotate 2
-    if (e.keyCode == key[4]) {
-      player.rotate(1);
+      if (e.type == 'keydown' && player.dropInterval != player.DROP_FAST) {
+        player.drop();
+        player.dropInterval = player.DROP_FAST;
+      } else {
+        player.dropInterval = player.DROP_SLOW;
+      }
     }
   });
-});
+}
+
+document.addEventListener('keydown', keyListeners);
+document.addEventListener('keyup', keyListeners);
